add functionality to sighup UI(auth):
A user service was created which a signup method. The
This service was used in the signup component.

Also, an alert service was created. It is used to pop up an alert when signup is successful. This service is available for use all through the applicationt display sever alert contents.

FINISHES SIGNUP FUNCTIONALITY