implements login(auth):
A method was created to handle signup in user service which returs an observable of error message when login is not successful vice versa
This method was used in login.component. In login component, after login is successful, user gets redirected to dashboard.

Note: no form of authorization was created in this commit. User can still access dashboard without login

FINISHES LOGIN FUNCTIONALITY