user authorization(auth)
A note service was created for fetching user's note. This service is neccessary at this stage because at this point of developement, the API in use with this application does not provide an easy way to get the currently logged in user. In view of this we eagerly load all user's notes. If there is a 401 error, it means user is not currently authenticated.

The note service was used in auth guard for user authorization. This application would use a proper user authorization mechanism as soon as the api in use provides a means to do that.

FINISHES USER AUTHORIZATION